refactor(donors): adopt @branch/lambda-http declarative route table
Replace the if-chain in handler.ts with a Route[] table dispatched via
@branch/lambda-http's dispatch(). handler.ts is now a one-liner; route
logic moved into controllers/donors.ts (GET/POST /donors, DELETE
/donors/:id) and controllers/donations.ts (GET/POST /donors/donations,
DELETE /donors/donations/:id), in the same order as the original if
chain. No services/ layer added, per this lambda's existing shape.

Local json() removed in favor of the shared one; auth stays manual
(authenticateRequest + custom 401/403 messages) since this lambda's
authorization messages don't match @branch/lambda-http's generic
requireAuth reasons. ROUTES-START/END markers moved into routes.ts,
now bracketing the route table entries.

Added @branch/lambda-http as a dependency and regenerated
package-lock.json. Added one test asserting GET /donors/donations
reaches the donations controller rather than a donor-id route.

No behavior change: same status codes, messages, and validation
order. Verified via tsc --noEmit and jest (--runInBand to avoid
DB contention with sibling lambda test runs): 51 passed, 1
pre-existing failure (health test requires a live dev-server on
:3000, fails identically on main).
