+++ apps/backend/lambdas/donors/package.json
@@ -24,6 +24,7 @@
   },
   "dependencies": {
     "@branch/lambda-auth": "file:../../../../shared/lambda-auth",
+    "@branch/lambda-http": "file:../../../../shared/lambda-http",
     "aws-jwt-verify": "^5.1.1",
     "kysely": "^0.28.8",
     "pg": "^8.17.2"
