+++ apps/backend/lambdas/donors/tsconfig.json
@@ -10,6 +10,6 @@
     "outDir": "dist",
     "sourceMap": true
   },
-  "include": ["*.ts", "jest.config.js"],
+  "include": ["*.ts", "controllers/**/*.ts", "jest.config.js"],
   "exclude": ["node_modules", "dist", "dev-server.ts", "swagger-utils.ts"]
 }
